refactored single component into seperate components / added util for goal / modal element with story and win modal
